Finally found an http client library
Finally found DavidWebb, an http client library that actually works in
java.

Finished implementing the barebones of the RobinhoodClient, it connects
to the website and logins in to obtain an authorization token. Moving
forward the access token will be used for navigating the account and
performing actions which require authorization.