switched back to other Robinhood API...
it did not go so well. The API is having a hard time contacting the
servers for some reason... I tried the latest OKHttp version of the api,
but it was throwing errors from GSON when I tried to log in with it,
then I changed it to the V1 version which came before that, and that I'm
fairly certain we were using before I started writing my client code,
and that is throwing errors from
conrad.weiser.robinhood.api.request.RequestManager.makePostRequest(RequestManager.java:115).

It says that there is some kind of error inside of the Robinhood API
library.