Started creating Little John Api
if we can get the other robinhood api working then this will just be a
sidekick api that will handle options functionality that the other
library is lacking.

If we can't get the other robinhood api working then this will take over
as the main robinhood api for the project.

Parker's TODO:
> LittleJohnApi
	a) log in / user info object
	b) fundamentals object
	c) instruments object
	d) stock options object
	e) exceptions objects for errors